Create module and fix linters errors
